Refactor of State class. Remove all DB logic from it.
