+++ Cm93.State/Interfaces/IRepository.cs
@@ -24,8 +24,9 @@ namespace Cm93.State.Interfaces
 {
 	public interface IRepository
 	{
-		void DeleteGame(Guid key);
+		void CreateGame(IState state);
+		void RetrieveGame(Guid key, IState state);
 		void UpdateGame(ModuleType moduleType, IState state);
-		IState LoadGame(Guid key);
+		void DeleteGame(Guid key);
 	}
 }
