+++ Cm93.UI/Cm93.UI.csproj
@@ -121,6 +121,7 @@
     </ApplicationDefinition>
     <Compile Include="Converters\ColourToBrushConverter.cs" />
     <Compile Include="CreateModel.cs" />
+    <Compile Include="Events\LoadGameEvent.cs" />
     <Compile Include="Events\MatchCompleteEvent.cs" />
     <Compile Include="Events\ModuleSelectedEvent.cs" />
     <Compile Include="Events\TeamSetEvent.cs" />
